+++ tests/test_meetup_to_apricot_event_adaptor.py
@@ -3,7 +3,8 @@
 from meetup2apricot.meetup_event import MeetupEvent
 from meetup2apricot.meetup_to_apricot_event_adaptor import MeetupToApricotEventAdaptor
 from .sample_apricot_json import EXPECTED_FREE_EVENT_JSON, \
-        EXPECTED_FREE_DESCRIPTION_HTML, EXPECTED_FREE_PHOTO_PATH
+        EXPECTED_FREE_DESCRIPTION_HTML, EXPECTED_FREE_PHOTO_PATH, \
+        EXPECTED_TAGS
 from datetime import datetime
 import pytest
 
@@ -11,12 +12,12 @@ import pytest
 @pytest.fixture()
 def free_event_adaptor(free_meetup_event_json):
     meetup_event = MeetupEvent(free_meetup_event_json)
-    return MeetupToApricotEventAdaptor(meetup_event, EXPECTED_FREE_PHOTO_PATH)
+    return MeetupToApricotEventAdaptor(meetup_event, EXPECTED_FREE_PHOTO_PATH, EXPECTED_TAGS)
 
 @pytest.fixture()
 def paid_event_adaptor(paid_meetup_event_json):
     meetup_event = MeetupEvent(paid_meetup_event_json)
-    return MeetupToApricotEventAdaptor(meetup_event, None)
+    return MeetupToApricotEventAdaptor(meetup_event, None, EXPECTED_TAGS)
 
 def test_name(free_event_adaptor):
     """Test returning the event title."""