+++ HISTORY.rst
@@ -15,6 +15,12 @@ Recent History
 .. Next Release
 .. ------------------
 
+0.1.1 (2020-11-08)
+------------------
+
+* Restrict event payments to online only.
+* Apply a configurable list of tags to all Wild Apricot events.
+
 0.1.0 (2020-11-08)
 ------------------